Updated usize -> u16 for recursion counter and limits

+++ massa-execution-exports/src/settings.rs
@@ -103,7 +103,7 @@ pub struct ExecutionConfig {
     /// Where to dump blocks
     pub block_dump_folder_path: PathBuf,
     /// Max recursive calls depth in SC
-    pub max_recursive_calls_depth: usize,
+    pub max_recursive_calls_depth: u16,
     /// Runtime condom middleware limits
     pub condom_limits: CondomLimits,
 }
